fix(market): order limit config fields were uint8 — 20000 wrapped to 32, truncating GetStationAsks to 32 types
THE root cause of 'нет в наличии' in the market item list. StationOrderLimit/
SystemOrderLimit/RegionOrderLimit (and the FindBuy/FindSell/OldPrice/NewPrice
limits) were declared uint8 in EVEServerConfig::market. Assigning 20000 to a
uint8 wraps to 32 (20000 % 256 = 32), so every ask query ran with LIMIT 32 —
GetStationAsks(60001795) returned a dict of 32 types instead of the ~10390
present in the DB. The client then showed 'нет в наличии' for every type not in
those 32. The compiler warning (unsigned conversion changes value 20000 to 32)
was there all along.

Widen all order-limit fields to uint32 (str2<uint32> and LIMIT %u both handle
it). OldPriceLimit/NewPriceLimit were also uint8: 1000 wrapped to 232.

+++ src/eve-server/EVEServerConfig.h
@@ -140,13 +140,13 @@ public:
     struct {
         bool UseOrderRange;
         bool DeleteOldTransactions;
-        uint8 StationOrderLimit;
-        uint8 SystemOrderLimit;
-        uint8 RegionOrderLimit;
-        uint8 FindBuyOrder;
-        uint8 FindSellOrder;
-        uint8 OldPriceLimit;
-        uint8 NewPriceLimit;
+        uint32 StationOrderLimit;
+        uint32 SystemOrderLimit;
+        uint32 RegionOrderLimit;
+        uint32 FindBuyOrder;
+        uint32 FindSellOrder;
+        uint32 OldPriceLimit;
+        uint32 NewPriceLimit;
         uint8 HistoryUpdateTime;
         float salesTax;
     } market;
